Test fix
I changed the tests to how i wrote them before in cest. i changed them to one function to see if the exceptions got throwed because there were to many functions. this wasnt it. i also stopped using _before and _after. with IE he opens a new window for every function and loads _before and _after for every function. this will give conflicts with the database that is why i made them myself.